[BE] Fix/person tag save review request dto birth year int to short

+++ BE/src/main/java/com/ssafy/matdongsan/domain/review/dto/PersonTagSaveReviewRequestDto.java
@@ -12,10 +12,10 @@ import lombok.NoArgsConstructor;
 @NoArgsConstructor(access = AccessLevel.PROTECTED)
 public class PersonTagSaveReviewRequestDto {
     private String name;
-    private int birthYear;
+    private short birthYear;
 
     @Builder
-    public PersonTagSaveReviewRequestDto(String name, int birthYear) {
+    public PersonTagSaveReviewRequestDto(String name, short birthYear) {
         this.name = name;
         this.birthYear = birthYear;
     }
